Particle system without texture + sTransformComponenet

+++ Engine/include/Engine/Graphics/Geometries/Plane.hpp
@@ -1,12 +1,10 @@
 #pragma once
 
-#include <cstdint>
-
 #include <Engine/Graphics/Model.hpp>
 
 class Plane: public Model {
 public:
     Plane();
-    Plane(uint32_t width, uint32_t height);
+    Plane(float width, float height);
     ~Plane();
 };
